requirements.txt
click==8.1.3
Flask==2.2.2
html2image==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.0
opencv-python==4.5.5.64
packaging==22.0
Pillow==9.3.0
pytesseract==0.3.10
python-dotenv==0.21.0
Werkzeug==2.2.2
